单容器部署 + 网页初始化向导（New API 式体验）+ SQLite 默认持久化
对齐 New API 的部署体验：一条 docker run + 浏览器网页初始化，零配置文件。

一、SQLite 默认持久化（架构级）
- 存储引擎三级：POSTGRES_DSN > SQLite(data/app.db) > 内存兜底
- 彻底解决审查报告"重启即失"：单机/单容器部署重启数据完整保留；
  多副本/大容量场景仍走 Postgres/Qdrant 编排
- 修复：MessageRow 主键 BigInteger→Integer（SQLite 自增别名）；
  _db_ready 自递归；e2e 测试前清库保证幂等（连跑两次 18/18）

二、网页初始化向导
- GET /auth/setup/status + POST /auth/setup（仅首次可用，此后永久 403）：
  网页创建平台管理员（旗舰档 + 同口令解锁 /admin 面板）
- 前端 /setup 向导页；登录页检测未初始化时显示引导入口

三、单容器镜像
- docker/single.Dockerfile：Node(前端 standalone) + Python(后端) 双运行时
  + ffmpeg/libmagic，SQLite 落 /app/data 卷
- docker run -d -p 3000:3000 -v studay-data:/app/data studay-rag:latest
- workflow 加第三个构建 job；compose.prod 内部服务（pg/redis/qdrant/minio）
  不再对外发布端口（仅 3000/8000）
- README：单容器部署文档

验证：setup 流程 6 项（首启判定/初始化/永久关闭/同口令面板/旗舰登录）+
e2e 连跑两次 18/18 + 单测 17 全过；tsc + build 通过，/setup 页入构建产物

+++ backend/app/db/pg_models.py
@@ -37,7 +37,8 @@ class UserRow(Base):
 class MessageRow(Base):
     __tablename__ = "messages"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    # Integer（非 BigInteger）：SQLite 仅对 INTEGER 主键启用自增别名
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     session_id: Mapped[str] = mapped_column(String(32), index=True)
     created_at: Mapped[int] = mapped_column(BigInteger, index=True, default=0)
     payload: Mapped[str] = mapped_column(Text)  # PolymorphicMessage JSON 全文
